### 2018-09-14 #942 @whedon
[PRE REVIEW]: hillR: taxonomic, functional, and phylogenetic diversity and similarity through Hill Numbers
**Submitting author:** @daijiang (<a href="http://orcid.org/0000-0002-0925-3421">Daijiang Li</a>)
**Repository:** <a href="https://github.com/daijiang/hillR" target ="_blank">https://github.com/daijiang/hillR</a>
**Version:** v0.3.0
