### 2018-09-14 #942 @aurielfournier
Hi @daijiang I don't do phylogenetic work, so I don't think I would be a very good reviewer for this package.. Thanks for thinking of me though
