### 2018-09-14 #942 @daijiang
Hi @aurielfournier , the methods have been proposed by others. This package implements it and it is actually straightforward. I think it is more about the R code than the methods. This is why I think you should be a good reviewer candidate.
