+++ R/hill_phylo.R
@@ -2,25 +2,26 @@
 #'
 #' @param comm data frame of vegtation data. Sites as rows, species as columns.
 #' @param tree a phylogeny with class 'phylo'.
+#' @author Chiu & Chao
 #' @noRd
 dat_prep_phylo <- function(comm, tree) {
-    if (class(tree) == "phylo") 
+    if (class(tree) == "phylo")
         tree <- ape::write.tree(tree)
     phyloData <- ade4::newick2phylog(tree)
     comm <- as.matrix(comm[, names(phyloData$leaves)])  # resort sp
     nodenames <- c(names(phyloData$leaves), names(phyloData$nodes))
-    
-    M <- matrix(0, nrow = ncol(comm), ncol = length(nodenames), dimnames = list(names(phyloData$leaves), 
+
+    M <- matrix(0, nrow = ncol(comm), ncol = length(nodenames), dimnames = list(names(phyloData$leaves),
         nodenames))
-    
+
     for (i in 1:nrow(M)) {
         M[i, ][unlist(phyloData$paths[i])] <- 1
     }
-    
+
     phylo_comm <- comm %*% M
     phyloLength <- c(phyloData$leaves, phyloData$nodes)
     treeH <- sum(phyloLength * phylo_comm[1, ]/sum(comm[1, ]))
-    
+
     return(list(pcomm = t(phylo_comm), pLength = phyloLength, treeH = treeH))
 }
 
@@ -34,7 +35,7 @@ dat_prep_phylo <- function(comm, tree) {
 #' @param base default is exp(1), the base of log.
 #' @param rel_then_pool default is TRUE. Abundance of species are changed to relative abundance within sites.
 #' @param show.warning whether to print warning, default is TRUE.
-#'
+#' @author Chiu & Chao
 #' @return a vector of hill number based phylogenetic diversity for all sites.
 #' @export
 #' @references Chao, Anne, Chun-Huo Chiu, and Lou Jost. Unifying Species Diversity, Phylogenetic Diversity, Functional Diversity, and Related Similarity and Differentiation Measures Through Hill Numbers. Annual Review of Ecology, Evolution, and Systematics 45, no. 1 (2014): 297–324. doi:10.1146/annurev-ecolsys-120213-091540.
@@ -47,47 +48,47 @@ dat_prep_phylo <- function(comm, tree) {
 #' hill_phylo(comm, tree, q = 2)
 #'
 hill_phylo <- function(comm, tree, q = 0, base = exp(1), rel_then_pool = TRUE, show.warning = TRUE) {
-    if (any(comm < 0)) 
+    if (any(comm < 0))
         stop("Negative value in comm data")
     # if(any(colSums(comm) == 0) & show.warning) warning('Some species in comm data were
     # not observed in any site,\n delete them...') comm = comm[, colSums(comm) != 0]
-    
+
     comm_sp <- intersect(colnames(comm), tree$tip.label)
-    
-    if (class(tree) != "phylo") 
+
+    if (class(tree) != "phylo")
         stop("tree must be an object with phylo as class")
     if (length(setdiff(tree$tip.label, comm_sp))) {
-        if (show.warning) 
+        if (show.warning)
             warning("Some species in the phylogeny but not in comm, \n remove them from the phylogeny...")
         tree <- ape::drop.tip(tree, tree$tip.label[!tree$tip.label %in% comm_sp])
     }
-    
+
     if (length(setdiff(colnames(comm), comm_sp))) {
-        if (show.warning) 
+        if (show.warning)
[…]
