+++ man/hill_phylo_parti.Rd
@@ -39,3 +39,6 @@ hill_phylo_parti(comm, tree, q = 2)
 \references{
 Chao, Anne, Chun-Huo Chiu, and Lou Jost. Unifying Species Diversity, Phylogenetic Diversity, Functional Diversity, and Related Similarity and Differentiation Measures Through Hill Numbers. Annual Review of Ecology, Evolution, and Systematics 45, no. 1 (2014): 297–324. doi:10.1146/annurev-ecolsys-120213-091540.
 }
+\author{
+Chiu & Chao, Daijiang Li
+}
