### 2018-11-08 #1041 @sydnerecord
> **Submitting author:** @daijiang ([Daijiang Li](http://orcid.org/0000-0002-0925-3421))
> **Repository:** https://github.com/daijiang/hillR
> **Version:** v0.3.0
> **Editor:** @karthik
> **Reviewer:** @sydnerecord
> **Archive:** Pending
>
> ## Status
> [![status](https://camo.githubusercontent.com/3346a78daec09aba01a470cc60435da135c007d7/687474703a2f2f6a6f73732e7468656f6a2e6f72672f7061706572732f35643361353163666336646362636462313030303435663863376463356266632f7374617475732e737667)](http://joss.theoj.org/papers/5d3a51cfc6dcbcdb100045f8c7dc5bfc)
>
> Status badge code:
>
> ```
> ```
> **Reviewers and authors**:
>
> Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the [target repository](https://github.com/daijiang/hillR) and link to those issues (especially acceptance-blockers) in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)
>
> ## Reviewer instructions & questions
> @sydnerecord, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
>
> 1. Make sure you're logged in to your GitHub account
> 2. Be sure to accept the invite at this URL: https://github.com/openjournals/joss-reviews/invitations
>
> The reviewer guidelines are available here: https://joss.theoj.org/about#reviewer_guidelines. Any questions/concerns please let @karthik know.
>
>  **Please try and complete your review in the next two weeks**
>
> ## Review checklist for @sydnerecord
> ### Conflict of interest
> * [x]  As the reviewer I confirm that I have read the [JOSS conflict of interest policy](https://github.com/openjournals/joss/blob/master/COI.md) and that there are no conflicts of interest for me to review this work.
>
> ### Code of Conduct
> * [x]  I confirm that I read and will adhere to the [JOSS code of conduct](https://joss.theoj.org/about#code_of_conduct).
>
> ### General checks
> * [x]  **Repository:** Is the source code for this software available at the [repository url](https://github.com/daijiang/hillR)?
> * [ ]  **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
> * [ ]  **Version:** Does the release version given match the GitHub release (v0.3.0)?
> * [ ]  **Authorship:** Has the submitting author (@daijiang) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
>
> ### Functionality
> * [ ]  **Installation:** Does installation proceed as outlined in the documentation?
> * [ ]  **Functionality:** Have the functional claims of the software been confirmed?
> * [ ]  **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
>
> ### Documentation
> * [ ]  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> * [ ]  **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
> * [ ]  **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
> * [ ]  **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
> * [ ]  **Automated tests:** Are there automated tests or manual steps described so that the function of the software can be verified?
> * [ ]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
> ### Software paper
> * [ ]  **Authors:** Does the `paper.md` file include a list of authors with their affiliations?
> * [ ]  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> * [ ]  **References:** Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
