### 2018-11-08 #1041 @sydnerecord
In the man files for hill_func.Rd, there is an argument {div_by_sp}{as FD calculated in this way will be highly correlated with taxonomic diversity, one simple way to correct this is to divide the results by the number of species. Default is FALSE.}
I'm not sure it it is really ok to simply correct by dividing the results by the number of species? Perhaps this issues requires a bit more context for the user. A common approach to dealing with this issue is to use null models to calculate standardized effect sizes (Swenson 2014). How would the specification of this argument in hillR intersect with this approach (or not)?
