+++ README.md
@@ -1,21 +1,25 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-[![Build Status](https://travis-ci.org/daijiang/hillR.svg?branch=master)](https://travis-ci.org/daijiang/hillR) [![Coverage status](https://codecov.io/gh/daijiang/hillR/branch/master/graph/badge.svg)](https://codecov.io/github/daijiang/hillR?branch=master)
 
-hillR
-=====
+[![Build
+Status](https://travis-ci.org/daijiang/hillR.svg?branch=master)](https://travis-ci.org/daijiang/hillR)
+[![Coverage
+status](https://codecov.io/gh/daijiang/hillR/branch/master/graph/badge.svg)](https://codecov.io/github/daijiang/hillR?branch=master)
 
-This package contains R functions to calculate taxonomic, functional, and phylogenetic diversity and site similarity through Hill Numbers. The underlying methods are based on Chao, Chiu and Jost 2014 and Chiu & Chao 2014.
+# hillR
 
-Installation
-============
+This package contains R functions to calculate taxonomic, functional,
+and phylogenetic diversity and site similarity through Hill Numbers. The
+underlying methods are based on Chao, Chiu and Jost 2014 and Chiu & Chao
+2014.
+
+# Installation
 
 To install this package, run the following code:
 
     devtools::install_github("daijiang/hillR")
 
-Examples
-========
+# Examples
 
 ``` r
 dummy = FD::dummy
@@ -25,8 +29,7 @@ tree = ape::rtree(n = ncol(comm), tip.label = paste0("sp", 1:ncol(comm)))
 library(hillR)
 ```
 
-Calculate taxonomic, functional, and phylogenetic diversity of each site
-------------------------------------------------------------------------
+## Calculate taxonomic, functional, and phylogenetic diversity of each site
 
 ``` r
 hill_taxa(comm, q = 0) # taxonomic alpha diversity
@@ -49,13 +52,12 @@ hill_func(comm, traits, q = 0) # functional alpha diversity
 
 hill_phylo(comm, tree, q = 0) # phylogenetic alpha diversity
 ##     com1     com2     com3     com4     com5     com6     com7     com8 
-## 4.607479 2.635823 3.372086 1.947757 3.490657 6.019674 3.539474 4.717898 
+## 5.320433 5.165530 4.209176 2.638789 4.190414 5.395492 2.885337 5.335720 
 ##     com9    com10 
-## 5.189872 4.381499
+## 4.557956 5.212277
 ```
 
-Calculate taxonomic, functional, and phylogenetic diversity across multiple sites
----------------------------------------------------------------------------------
+## Calculate taxonomic, functional, and phylogenetic diversity across multiple sites
 
 ``` r
 hill_taxa_parti(comm, q = 0) # taxonomic diversity across all sites
@@ -70,27 +72,26 @@ hill_func_parti(comm, traits, q = 0) # functional diversity across all sites
 
 hill_phylo_parti(comm, tree, q = 0) # phylogenetic diversity across all sites
 ##   q PD_gamma PD_alpha  PD_beta local_similarity region_similarity
-## 1 0 6.602742 3.990222 1.654731        0.9272522         0.5603644
+## 1 0 7.359797 4.491113 1.638747        0.9290281         0.5669137
 ```
 
-Calculate pairwise taxonomic, functional, and phylogenetic diversity
---------------------------------------------------------------------
+## Calculate pairwise taxonomic, functional, and phylogenetic diversity
 
 ``` r
 hill_taxa_parti_pairwise(comm, q = 0, show.warning = F) # pairwise taxonomic diversity
 ## # A tibble: 45 x 8
[…]
