+++ joss/paper.bib
@@ -5,7 +5,8 @@
   volume={45},
   pages={297--324},
   year={2014},
-  publisher={Annual Reviews}
+  publisher={Annual Reviews},
+  doi={10.1146/annurev-ecolsys-120213-091540}
 }
 
 @article{chiu2014distance,
@@ -16,7 +17,8 @@
   number={7},
   pages={e100014},
   year={2014},
-  publisher={Public Library of Science}
+  publisher={Public Library of Science},
+  doi={10.1371/journal.pone.0100014}
 }
 
 @article{hill1973diversity,
@@ -27,7 +29,8 @@
   number={2},
   pages={427--432},
   year={1973},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.2307/1934352}
 }
 
 @article{jost2006entropy,
@@ -38,5 +41,6 @@
   number={2},
   pages={363--375},
   year={2006},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/j.2006.0030-1299.14714.x}
 }
