+++ R/hill_func.R
@@ -16,9 +16,9 @@
 #' @param fdis whether to calculated FDis, default is TRUE.
 #' @param stand_dij whether to standardize distance matrix to have max value of 1? Default is FALSE.
 #' @export
-#' @references Chao, Anne, Chun-Huo Chiu, and Lou Jost. Unifying Species Diversity, Phylogenetic Diversity, Functional Diversity, and Related Similarity and Differentiation Measures Through Hill Numbers. Annual Review of Ecology, Evolution, and Systematics 45, no. 1 (2014): 297–324. doi:10.1146/annurev-ecolsys-120213-091540.
+#' @references Chao, Anne, Chun-Huo Chiu, and Lou Jost. Unifying Species Diversity, Phylogenetic Diversity, Functional Diversity, and Related Similarity and Differentiation Measures Through Hill Numbers. Annual Review of Ecology, Evolution, and Systematics 45, no. 1 (2014): 297–324. <doi:10.1146/annurev-ecolsys-120213-091540>.
 #'
-#' Chiu, Chun-Huo, and Anne Chao. Distance-Based Functional Diversity Measures and Their Decomposition: A Framework Based on Hill Numbers. PLoS ONE 9, no. 7 (July 7, 2014): e100014. doi:10.1371/journal.pone.0100014.
+#' Chiu, Chun-Huo, and Anne Chao. Distance-Based Functional Diversity Measures and Their Decomposition: A Framework Based on Hill Numbers. PLoS ONE 9, no. 7 (July 7, 2014): e100014. <doi:10.1371/journal.pone.0100014>.
 #' @return a matrix, with these information for each site: Q (Rao's Q); D_q (functional hill number,
 #'  the effective number of equally abundant and functionally equally distince species);
 #'  MD_q (mean functional diversity per species, the effective sum of pairwise distances between
@@ -33,10 +33,10 @@
 #' hill_func(comm = dummy$abun, traits = dummy$trait, q = 2)
 #' hill_func(comm = dummy$abun, traits = dummy$trait, q = 3)
 #'
-hill_func <- function(comm, traits, traits_as_is = FALSE, q = 0, base = exp(1), checkdata = TRUE, 
+hill_func <- function(comm, traits, traits_as_is = FALSE, q = 0, base = exp(1), checkdata = TRUE,
     div_by_sp = FALSE, ord = c("podani", "metric"), fdis = TRUE, stand_dij = FALSE) {
     if (checkdata) {
-        if (any(comm < 0)) 
+        if (any(comm < 0))
             stop("Negative value in comm data")
         if (is.null(rownames(traits))) {
             stop("\n Traits have no row names\n")
@@ -45,14 +45,14 @@ hill_func <- function(comm, traits, traits_as_is = FALSE, q = 0, base = exp(1), 
             stop("\n Comm data have no col names\n")
         }
     }
-    
+
     if (any(!colnames(comm) %in% rownames(traits))) {
         warning("\n There are species from community data that are not on traits matrix\nDelete these species from comm data...\n")
         comm <- comm[, colnames(comm) %in% rownames(traits)]
     }
-    
+
     # all(rownames(traits) == names(comm))
-    
+
     if (traits_as_is) {
         # traits is already a distance matrix
         dij <- as.matrix(traits)
@@ -63,11 +63,11 @@ hill_func <- function(comm, traits, traits_as_is = FALSE, q = 0, base = exp(1), 
         traits <- plyr::arrange(traits[traits$sp %in% colnames(comm), ], sp)
         rownames(traits) <- traits$sp
[…]
