+++ R/hill_phylo.R
@@ -38,7 +38,7 @@ dat_prep_phylo <- function(comm, tree) {
 #' @author Chiu & Chao
 #' @return a vector of hill number based phylogenetic diversity for all sites.
 #' @export
-#' @references Chao, Anne, Chun-Huo Chiu, and Lou Jost. Unifying Species Diversity, Phylogenetic Diversity, Functional Diversity, and Related Similarity and Differentiation Measures Through Hill Numbers. Annual Review of Ecology, Evolution, and Systematics 45, no. 1 (2014): 297–324. doi:10.1146/annurev-ecolsys-120213-091540.
+#' @references Chao, Anne, Chun-Huo Chiu, and Lou Jost. Unifying Species Diversity, Phylogenetic Diversity, Functional Diversity, and Related Similarity and Differentiation Measures Through Hill Numbers. Annual Review of Ecology, Evolution, and Systematics 45, no. 1 (2014): 297–324. <doi:10.1146/annurev-ecolsys-120213-091540>.
 #' @examples
 #' comm = dummy = FD::dummy$abun
 #' tree = ape::rtree(n = ncol(comm), tip.label = paste0('sp', 1:8))

[…]
