### 2018-11-08 659c9f7
update div_by_sp #6
