### 2018-11-08 #1041 @daijiang
> The license is an MIT license in a LICENSE.md file. I have not reviewed for your journal before, and I'm not sure if this meets the criteria.

The reason for this is because of CRAN's license policy...
