### 2018-11-08 #1041 @karthik
@sydnerecord MIT is an approved OSI license (https://opensource.org/licenses/alphabetical) so that is perfectly ok here. For specific issues with the software, (in future) please open issues on the submitted repo and add `openjournals/joss-reviews#1041`  to crosslink here.
